requirements.txt
# StarRocks 慢 SQL 分析系统依赖
pymysql==1.1.0
starrocks==1.0.0
pandas==2.1.4
sqlparse==0.4.4
jinja2==3.1.2
python-dateutil==2.8.2
tabulate==0.9.0
colorama==0.4.6
pydantic==2.5.3
pyyaml==6.0.1

